refs #192 - move Faqs & Faq to their own folder
