+++ src/components/pricing/faqs.svelte
@@ -1,8 +1,8 @@
 <script>
   import { isEurope } from "../../utils/helper";
   import Section from "../section.svelte";
-  import FaqsContainer from "./faqs-container.svelte";
-  import Faq from "./faq.svelte";
+  import Faqs from "../faqs/faqs.svelte";
+  import Faq from "../faqs/faq.svelte";
 </script>
 
 <style lang="scss">
@@ -21,7 +21,7 @@
 </style>
 
 <Section>
-  <FaqsContainer>
+  <Faqs>
     <h2 class="h1">FAQs</h2>
     <Faq title="Can I always use Gitpod for free?">
       <p>
@@ -94,5 +94,5 @@
         We are happy to answer them, please <a href="/contact">Get in Touch</a>.
       </p>
     </Faq>
-  </FaqsContainer>
+  </Faqs>
 </Section>
